requirements.txt
ansible==2.8.20
appdirs==1.4.4
apturl==0.5.2
astroid==2.4.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
click==8.0.3
colorama==0.4.3
command-not-found==0.3
crcmod==1.7
cryptography==3.4.7
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==2.0.1
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
linecache2==1.0.0
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.20.3
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
packaging==20.9
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
pluggy==0.13.1
protobuf==3.6.1
py==1.10.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.5.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
PyPDF2==1.26.0
pyRFC3339==1.1
pytest==6.2.4
pytest-json-report==1.4.0
pytest-metadata==1.11.0
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
SecretStorage==2.3.1
semantic-version==2.8.5
setuptools-rust==0.12.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
stevedore==1.32.0
systemd-python==234
testresources==2.0.0
toml==0.10.0
traceback2==1.4.0
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unittest2==1.1.0
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wadllib==1.3.3
Werkzeug==2.0.2
wrapt==1.12.1
xkit==0.0.0
zipp==1.0.0
